refactor: isolate child task context assembly
