+++ crates/agent-engine/src/runtime/child_agents/task_context.rs
@@ -1,0 +1,153 @@
+use crate::runtime::agent_loop::RuntimeLoopState;
+use crate::tape::Message;
+use alan_agent_protocol::{SpawnHandle, SpawnSpec};
+
+const MAX_CHILD_CONVERSATION_MESSAGES: usize = 8;
+const MAX_CHILD_CONVERSATION_CHARS: usize = 4_000;
+const MAX_CHILD_PLAN_ITEMS: usize = 16;
+const MAX_CHILD_PLAN_ITEM_CHARS: usize = 240;
+const MAX_CHILD_TOOL_RESULTS: usize = 6;
+const MAX_CHILD_TOOL_RESULT_CHARS: usize = 1_200;
+
+pub(super) fn build_child_task_text(parent: &RuntimeLoopState, spec: &SpawnSpec) -> String {
+    let mut sections = vec![spec.launch.task.trim().to_string()];
+
+    if let Some(metadata) = render_launch_metadata(spec) {
+        sections.push(metadata);
+    }
+    if spec.has_handle(SpawnHandle::ConversationSnapshot)
+        && let Some(snapshot) = render_conversation_snapshot(parent)
+    {
+        sections.push(snapshot);
+    }
+    if spec.has_handle(SpawnHandle::Plan)
+        && let Some(snapshot) = render_plan_snapshot(parent)
+    {
+        sections.push(snapshot);
+    }
+    if spec.has_handle(SpawnHandle::ToolResults)
+        && let Some(snapshot) = render_tool_results_snapshot(parent)
+    {
+        sections.push(snapshot);
+    }
+
+    sections
+        .into_iter()
+        .filter(|section| !section.trim().is_empty())
+        .collect::<Vec<_>>()
+        .join("\n\n")
+}
+
+fn render_launch_metadata(spec: &SpawnSpec) -> Option<String> {
+    let mut lines = Vec::new();
+    if let Some(cwd) = spec.launch.cwd.as_ref() {
+        lines.push(format!("cwd: {}", cwd.display()));
+    }
+    if let Some(output_dir) = spec.launch.output_dir.as_ref() {
+        lines.push(format!("output_dir: {}", output_dir.display()));
+    }
+
+    (!lines.is_empty()).then(|| format!("Execution Context\n{}", lines.join("\n")))
+}
+
+fn render_conversation_snapshot(parent: &RuntimeLoopState) -> Option<String> {
+    let mut lines = Vec::new();
+    if let Some(summary) = parent.machine.tape.summary() {
+        lines.push("Summary:".to_string());
+        lines.push(truncate_chars(summary.trim(), MAX_CHILD_CONVERSATION_CHARS));
+    }
+
+    let recent_messages = parent
+        .machine
+        .tape
+        .messages()
+        .iter()
+        .rev()
+        .filter(|message| matches!(message, Message::User { .. } | Message::Assistant { .. }))
+        .take(MAX_CHILD_CONVERSATION_MESSAGES)
+        .cloned()
+        .collect::<Vec<_>>();
+
+    if !recent_messages.is_empty() {
+        lines.push("Recent Messages:".to_string());
+        for message in recent_messages.into_iter().rev() {
+            let role = match &message {
+                Message::User { .. } => "user",
+                Message::Assistant { .. } => "assistant",
+                Message::Tool { .. } => unreachable!("tool messages are filtered out above"),
+                Message::System { .. } => "system",
+                Message::Context { .. } => "context",
+            };
+            let text = match &message {
+                Message::Assistant { .. } => message.non_thinking_text_content(),
+                _ => message.text_content(),
+            };
+            if !text.trim().is_empty() {
+                lines.push(format!(
+                    "- {role}: {}",
+                    truncate_chars(text.trim(), MAX_CHILD_CONVERSATION_CHARS / 2)
+                ));
+            }
+        }
+    }
+
+    (!lines.is_empty()).then(|| format!("Parent Conversation Snapshot\n{}", lines.join("\n")))
+}
+
+fn render_plan_snapshot(parent: &RuntimeLoopState) -> Option<String> {
+    let plan_snapshot = parent.turn_state.plan_snapshot()?;
+    let mut lines = Vec::new();
+    if let Some(explanation) = plan_snapshot.explanation.as_deref()
+        && !explanation.trim().is_empty()
+    {
+        lines.push(format!(
+            "Explanation: {}",
+            truncate_chars(explanation.trim(), MAX_CHILD_PLAN_ITEM_CHARS)
+        ));
+    }
+    for item in plan_snapshot.items.iter().take(MAX_CHILD_PLAN_ITEMS) {
+        lines.push(format!(
+            "- [{}] {}",
+            match item.status {
+                alan_agent_protocol::PlanItemStatus::Pending => "pending",
+                alan_agent_protocol::PlanItemStatus::InProgress => "in_progress",
+                alan_agent_protocol::PlanItemStatus::Completed => "completed",
+            },
+            truncate_chars(item.content.trim(), MAX_CHILD_PLAN_ITEM_CHARS)
+        ));
+    }
+
+    (!lines.is_empty()).then(|| format!("Parent Plan Snapshot\n{}", lines.join("\n")))
+}
+
+fn render_tool_results_snapshot(parent: &RuntimeLoopState) -> Option<String> {
+    let mut lines = Vec::new();
+    for message in parent
+        .machine
+        .tape
+        .messages()
+        .iter()
+        .rev()
+        .filter(|message| matches!(message, Message::Tool { .. }))
+        .take(MAX_CHILD_TOOL_RESULTS)
+    {
+        for response in message.tool_responses() {
+            let content =
+                truncate_chars(response.text_content().trim(), MAX_CHILD_TOOL_RESULT_CHARS);
+            if !content.is_empty() {
+                lines.push(format!("- {}: {}", response.id, content));
+            }
+        }
+    }
+    lines.reverse();
+    (!lines.is_empty()).then(|| format!("Parent Tool Results\n{}", lines.join("\n")))
+}
+
+fn truncate_chars(text: &str, limit: usize) -> String {
+    let truncated: String = text.chars().take(limit).collect();
+    if truncated.chars().count() == text.chars().count() {
+        truncated
+    } else {
+        format!("{truncated}...")
+    }
+}
